Added better position printing
Makes it so that instead of printing line and character number
we print piece of code specified by position of an error

Before:
```
$ dbl test/err/tc_0005_specialBinops.dbl
test/err/tc_0005_specialBinops.dbl:1:9-10: error: This expression has
type String, but an expression was expected of type Bool
test/err/tc_0005_specialBinops.dbl:2:9-10: error: This expression has
type String, but an expression was expected of type Bool
test/err/tc_0005_specialBinops.dbl:2:15-16: error: This expression has
type String, but an expression was expected of type Bool
test/err/tc_0005_specialBinops.dbl:2:21-22: error: This expression has
type String, but an expression was expected of type Bool
test/err/tc_0005_specialBinops.dbl:3:17: error: This expression has type
Int, but an expression was expected of type Bool
test/err/tc_0005_specialBinops.dbl:4:9-10: error: This expression has
type String, but an expression was expected of type Unit
```

Now:
```
$ dbl test/err/tc_0005_specialBinops.dbl
error: This expression has type String, but an expression was expected of type Bool
-> test/err/tc_0005_specialBinops.dbl
 1 | let _ = "" && True
   |         ^^
error: This expression has type String, but an expression was expected of type Bool
-> test/err/tc_0005_specialBinops.dbl
 2 | let _ = "" && "" || ""
   |         ^^
error: This expression has type String, but an expression was expected of type Bool
-> test/err/tc_0005_specialBinops.dbl
 2 | let _ = "" && "" || ""
   |               ^^
error: This expression has type String, but an expression was expected of type Bool
-> test/err/tc_0005_specialBinops.dbl
 2 | let _ = "" && "" || ""
   |                     ^^
error: This expression has type Int, but an expression was expected of type Bool
-> test/err/tc_0005_specialBinops.dbl
 3 | let _ = True || 5
   |                 ^
error: This expression has type String, but an expression was expected of type Unit
-> test/err/tc_0005_specialBinops.dbl
 4 | let _ = ""; 5
   |         ^^
```

I think this is a good quality of life improvement.

It's possible that after this change we will detect that somewhere
position isn't combined, or passed correctly.