Fixes proposed in review
For me, it looked like moving just external functions tbl to another file
would be such a mess because it uses `type value` so that would need
separate file, so moving whole thing to another subdirectory was the best
solution

As to not having to do `import Int64` just to say `2L + 2L`, i think
having sub modules in Prelude is a good idea and maybe doing it with
other builtin types is as well. Especially since methods can be closed
behind module and are still usable as long as module is public

I don't a problem in having both (int).toInt64 and Int64.ofInt in stdlib
especially if one uses the other.