Tree-like representation of scopes
This approach simplifies data structures used to represent scopes of
unification variables, and allows to get rid of permutations. Obtained
type-reconstruction algorithm is slightly weaker, but more predictable:
unification variables under the binder of a type variable always doesn't
have this variable in scope (for the previous algorithm it usually was
the case, but not always), and therefore cannot be set to something that
contains this variable.