Addressing the comments
Actually, the commit slightly changes the approach to default printing
method. Instead of using flag (or `option` type), the printing function
is built during the type inference phase. It uses `toString` method if
it is available, otherwise it uses `dbl_toString` extern that is
implemented as the old `to_string` function of the evaluator.