Relaxed method resolution, based on sizes
This commit proposes a relaxed restriction of a method resolution. Instead
of forbidding using the same method twice, it allows multiple use of the
same method, but each consecutive usage must be on the smaller type.
Since unification variables have size 1, the solution is not perfect, but
allows, e.g., for pretty-printing of nested lists.