Detecting and simplifying equality-like constraints (#282)
This commit introduces a new simplification rule for subeffecting
constraints: if the equality between generalizable variable and some
effect can be deduced from the set of constraints, the variable is set
to that effect. The algorithm doesn't look very efficient, but with
this rule, type schemes contains virtually no constraints in real
programs, so the effect inference works really fast.